add mean_all optest
